add improved code and variables module
