+++ js/my_profile.mjs
@@ -1,3 +1,3 @@
-const myName = localStorage.getItem("userName");
+import { myName } from './variables.mjs';
 
 window.location.href = `/pages/profile.html?user=${myName}`;
