+++ pages/log_in.html
@@ -9,10 +9,10 @@
     <link rel="stylesheet" href="/css/styles.css">
     <script src="../node_modules/bootstrap/dist/js/bootstrap.bundle.min.js" defer></script>
     <script type="module" src="../js/variables.mjs"></script>
-    <script src="../js/searchbar.js" defer></script>
-    <script src="../js/login.js" defer></script>
-    <script src="../js/logged_in.js" defer></script>
-    <script src="../js/create_post_mobile.js" defer></script>
+    <script type="module" src="../js/searchbar.mjs"></script>
+    <script type="module" src="../js/login.mjs"></script>
+    <script type="module" src="../js/logged_in.mjs"></script>
+    <script type="module" src="../js/create_post_mobile.mjs"></script>
 </head>
 <body>
     <!-- header -->
