+++ pages/search.html
@@ -9,12 +9,12 @@
     <link rel="stylesheet" href="/css/styles.css">
     <script src="../node_modules/bootstrap/dist/js/bootstrap.bundle.min.js" defer></script>
     <script type="module" src="../js/variables.mjs"></script>
-    <script src="../js/searchbar.js" defer></script>
-    <script src="../js/search.js" defer></script>
+    <script type="module" src="../js/searchbar.mjs"></script>
+    <script type="module" src="../js/search.mjs"></script>
     <script src="../js/login_check.js" defer></script>
-    <script src="../js/logged_in.js" defer></script>
+    <script type="module" src="../js/logged_in.mjs"></script>
     <script type="module" src="../js/create_post.mjs"></script>
-    <script src="../js/create_post_mobile.js" defer></script>
+    <script type="module" src="../js/create_post_mobile.mjs"></script>
 </head>
 <body>
     <!-- header -->
